refacto: simplify ssr static props example
