+++ package.json
@@ -12,6 +12,7 @@
     "build:watch": "pnpm --parallel --color --filter '@wbe/*' run build:watch",
     "dev": "pnpm --parallel --color --filter './packages/**/examples/*' run dev",
     "dev:example-preact-ssr": "pnpm --filter example-preact-ssr dev",
+    "dev:example-preact-ssr-static-props": "pnpm --filter example-preact-ssr-static-props dev",
     "test:watch": "vitest --reporter verbose",
     "test": "vitest run",
     "size": "size-limit",
