--- packages/low-router-preact/examples/example-preact-ssr-static-props/.env
@@ -1,1 +1,0 @@
-VITE_APP_BASE=/
