--- packages/low-router-preact/examples/example-preact-ssr-static-props/config/config.js
@@ -1,7 +1,0 @@
-import { resolve } from "path"
-
-export default {
-  srcDir: resolve("src"),
-  publicDir: resolve("src/public"),
-  outDir: resolve("dist"),
-}
