--- packages/low-router-preact/examples/example-preact-ssr-static-props/index.html
@@ -1,16 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="x-ua-compatible" content="IE=Edge" />
-    <meta
-      name="viewport"
-      content="width=device-width, initial-scale=1, maximum-scale=1"
-    />
-    <title>SPA APP</title>
-  </head>
-  <body>
-    <div id="root"></div>
-    <script type="module" src="/src/index-client.tsx"></script>
-  </body>
-</html>
