+++ packages/low-router-preact/examples/example-preact-ssr-static-props/src/App.tsx
@@ -1,17 +1,13 @@
-import css from "./App.module.scss"
-import { Link, Stack, useRouter } from "@wbe/low-router-preact"
+import { Link, Stack } from "@wbe/low-router-preact"
 
 function App() {
-  const router = useRouter()
-
   return (
-    <div className={css.root}>
+    <div>
       <div>
         <Link to={{ name: "home" }}>{"home"}</Link>
         <Link to={{ name: "work", params: { id: "id-1" } }}>{"work id-1"}</Link>
         <Link to={{ name: "work", params: { id: "id-2" } }}>{"work id-2"}</Link>
       </div>
-
       <Stack clampRoutesRender={false} />
     </div>
   )
