--- packages/low-router-preact/examples/example-preact-ssr-static-props/src/components/App/App.module.scss
@@ -1,6 +1,0 @@
-.root {
-
-  a {
-    display: block;
-  }
-}
