--- packages/low-router-preact/examples/example-preact-ssr-static-props/src/core/server-utils/RawScript.tsx
@@ -1,16 +1,0 @@
-/**
- * Insert raw script in window variable
- * @param name
- * @param obj
- */
-export const RawScript = ({ name, data }) => {
-  const stringify = (e): string => JSON.stringify(e, null, 2)?.replace(/\n\s+/g, "")
-  return (
-    <script
-      type="text/javascript"
-      dangerouslySetInnerHTML={{
-        __html: `window.${name}=${stringify(data)}`,
-      }}
-    />
-  )
-}
