+++ packages/low-router-preact/examples/example-preact-ssr-static-props/src/index-client.tsx
@@ -1,20 +1,19 @@
 import "preact/debug"
 import "./index.scss"
-import App from "~/src/components/App/App"
+import App from "~/src/App"
 import { hydrate } from "preact"
-import { locales, routes, defaultLocaleInUrl } from "~/src/routes"
+import { routes } from "~/src/routes"
 import { createBrowserHistory, LowRouter } from "@wbe/low-router"
 import { StrictMode } from "preact/compat"
-import { I18n, Router } from "@wbe/low-router-preact"
+import { Router } from "@wbe/low-router-preact"
 
 const base: string = import.meta.env.VITE_APP_BASE || "/"
-const i18n = new I18n(locales, { base, defaultLocaleInUrl })
-const router = new LowRouter(i18n.addLocaleParamToRoutes(routes), { base, id: 1 })
+
+const router = new LowRouter(routes, { base, id: 1 })
 
 hydrate(
   <StrictMode>
     <Router
-      i18n={i18n}
       router={router}
       history={createBrowserHistory()}
       initialStaticProps={window["__INITIAL_STATIC_PROPS__"]}
