+++ packages/low-router-preact/examples/example-preact-ssr-static-props/src/index.scss
@@ -14,4 +14,5 @@ body {
 
 a {
   color: #3aa48f;
+  display: block;
 }
